Fonctionne mais pb d'envoi de température du core
Température du core transmise est de -180 deg

+++ humiditySensor/humiditySensor/Message.h
@@ -4,7 +4,7 @@ class Message
 {
 public:
 	Message();
-	void encodeAndSendOneMessage(const char type, int sensorID, double data);
+	void encodeAndSendOneMessage(const char type, int sensorID, int data);
 	int calculatePayloadLenght(int value);
 	void choseTxAndRxPipe(void);
 	void transmitterInit();
